# govconnect-dashboard/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GovConnect API
  description: |
    Complete API documentation for GovConnect microservices platform.
    
    ## Services
    - **Case Service** - Complaints & Service Requests management
    - **Channel Service** - WhatsApp Channel & Live Chat
    - **AI Service** - AI Orchestrator & Knowledge Base
    - **Notification Service** - Event-driven notifications
    
    ## Authentication
    - **Internal API Key**: `X-Internal-API-Key` header for service-to-service communication
    - **JWT Token**: `Authorization: Bearer <token>` for dashboard authentication
    - **Public APIs**: Endpoint publik tersedia untuk form publik (tanpa auth)
    
    ## Base URLs
    Layanan diakses langsung per service (tanpa gateway agregasi).
  version: 1.0.0
  contact:
    name: GovConnect Team
    url: https://govconnect.my.id
  license:
    name: MIT
servers:
  - url: https://case.govconnect.my.id
    description: Case Service (Production)
  - url: https://channel.govconnect.my.id
    description: Channel Service (Production)
  - url: https://ai.govconnect.my.id
    description: AI Service (Production)
...
components:
  securitySchemes:
    InternalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
      description: Internal API key for service-to-service communication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for dashboard authentication
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
        wa_user_id:
          type: string
        message:
          type: string
        direction:
          type: string
          enum: [IN, OUT]
        source:
          type: string
          enum: [USER, AI, ADMIN]
        created_at:
          type: string
          format: date-time
...
paths:
...
  /internal/messages:
    get:
      tags:
        - Messages
      summary: Get messages
      description: Get messages for a user
      operationId: getMessages
      security:
        - InternalApiKey: []
      parameters:
        - name: wa_user_id
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
    post:
      tags:
        - Messages
      summary: Store message
      description: Store AI reply in database
      operationId: storeMessage
      security:
        - InternalApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - wa_user_id
                - message
                - direction
              properties:
                wa_user_id:
                  type: string
                message:
                  type: string
                direction:
                  type: string
                  enum: [IN, OUT]
                source:
                  type: string
                  enum: [USER, AI, ADMIN]
      responses:
        '201':
          description: Message stored
...
  /internal/messages/read:
    post:
      tags:
        - Messages
      summary: Mark messages as read
      operationId: markMessagesRead
      security:
        - InternalApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - wa_user_id
              properties:
                wa_user_id:
                  type: string
                message_ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Messages marked as read

  # ============================================
  # CHANNEL SERVICE - WhatsApp Session
  # ============================================
...